+++ app/lib/fake/backend/fake_dartdoc_runner.dart
@@ -12,12 +12,6 @@ import 'package:pub_dartdoc_data/pub_dartdoc_data.dart';
 
 import '../../dartdoc/dartdoc_runner.dart';
 
-/// Generates package documentation for all packages with fake dartdoc runner.
-Future<void> processJobsWithFakeDartdocRunner() async {
-  // ignore: invalid_use_of_visible_for_testing_member
-  await processJobsWithDartdocRunner(runner: FakeDartdocRunner());
-}
-
 /// Generates dartdoc content and results based on a deterministic random seed.
 class FakeDartdocRunner implements DartdocRunner {
   @override