+++ app/test/service/security_advisory/security_advisory_test.dart
@@ -1,15 +1,71 @@
-// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
+import 'dart:io';
 
 import 'package:clock/clock.dart';
+import 'package:path/path.dart' as path;
 import 'package:pub_dev/service/security_advisories/backend.dart';
 import 'package:pub_dev/service/security_advisories/models.dart';
+import 'package:pub_dev/shared/utils.dart';
 import 'package:test/test.dart';
 
-import '../shared/test_services.dart';
+import '../../shared/test_services.dart';
 
 void main() {
+  test('Parse advisory', () async {
+    try {
+      final file = File(path.join(Directory.current.path, 'test', 'service',
+              'security_advisory', 'testdata', 'example_advisory.json'))
+          .readAsBytesSync();
+      OSV.fromJson(utf8JsonDecoder.convert(file) as Map<String, dynamic>);
+    } catch (e, st) {
+      fail('Parsing advisory failed with \n$e\n$st');
+    }
+  });
+
+  testWithProfile('List all advisories and delete advisory', fn: () async {
+    final firstTime = DateTime(2022).toIso8601String();
+    final affectedA = Affected(
+      package: Package(ecosystem: 'pub', name: 'a'),
+      versions: ['1'],
+    );
+    final id = '123';
+    final id2 = '456';
+
+    final osv = OSV(
+      schemaVersion: '1.2.3',
+      id: id,
+      modified: firstTime,
+      published: firstTime,
+      affected: [affectedA],
+    );
+
+    final osv2 = OSV(
+      schemaVersion: '1.2.3',
+      id: id2,
+      modified: firstTime,
+      published: firstTime,
+      affected: [affectedA],
+    );
+
+    await securityAdvisoryBackend.ingestSecurityAdvisory(osv);
+    await securityAdvisoryBackend.ingestSecurityAdvisory(osv2);
+
+    final all = await securityAdvisoryBackend.listAdvisories();
+    expect(all, isNotNull);
+    expect(all.length, 2);
+    expect(all.first.id, id);
+    expect(all.last.id, id2);
+
+    await securityAdvisoryBackend.deleteAdvisory(id);
+
+    final reduced = await securityAdvisoryBackend.listAdvisories();
+    expect(reduced, isNotNull);
+    expect(reduced.length, 1);
+    expect(reduced.first.id, id2);
+
+    final advisory = await securityAdvisoryBackend.lookupById(id);
+    expect(advisory, isNull);
+  });
+
   testWithProfile('Insert, lookup and update advisory', fn: () async {
     final firstTime = DateTime(2022).toIso8601String();
     final affectedA = Affected(